+++ .eslintrc.js
@@ -18,6 +18,7 @@ const config = {
   rules: {
     'no-console': 'off',
     'no-underscore-dangle': 'off',
+    camelcase: 'off',
 
     'import/no-extraneous-dependencies': 'off',
     'import/no-unresolved': 'off',
@@ -71,7 +72,8 @@ const config = {
     '@typescript-eslint/class-name-casing': 'off',
     '@typescript-eslint/no-unused-vars': 'off',
     '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/ban-ts-ignore': 'off',
+    '@typescript-eslint/ban-ts-comment': 'off',
+    '@typescript-eslint/explicit-module-boundary-types': 'off',
 
     'jest/no-expect-resolves': 'error'
   }